requirements.txt
fastapi==0.115.1 
uvicorn==0.34.0 
mysql-connector-python==9.2.0 
SQLAlchemy==1.4.0 
google-generativeai==0.8.4 
